Move noDuplicateUnderscoresRecursive and testing to helper function files
